+++ db/migrate/20200421223602_add_status_to_integral_users.rb
@@ -1,0 +1,5 @@
+class AddStatusToIntegralUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :integral_users, :status, :integer, default: 0
+  end
+end